v0.9

+++ delete_post.php
@@ -16,8 +16,11 @@ $response_message='';
             $sql="SELECT image FROM posts WHERE post_id=".$_POST['post_id'];
             $data=mysqli_query($con, $sql);
             $data=mysqli_fetch_all($data,MYSQLI_ASSOC);
-            $delfile=$_SERVER['DOCUMENT_ROOT'].'/posts_images/'.$data[0]['image'];
-           if (file_exists($delfile)) { unlink ($delfile); }
+            if(isset($data[0]['image'])&& strlen($data[0]['image'])>0){
+             $delfile=$_SERVER['DOCUMENT_ROOT'].'/posts_images/'.$data[0]['image'];
+                if (file_exists($delfile)) { unlink ($delfile); }
+            }
+           
             $sql="DELETE FROM posts WHERE post_id=".$_POST['post_id'];
             if ($con->query($sql) === TRUE) {
                 $response_message= 'Post deleted succesfully';
